+++ web/static/react/assets/PracticeProfilePage-DuvXa-wV.js
@@ -1,1 +1,0 @@
-import{j as r}from"./vendor-react-DCA3L5Xq.js";import{L as o}from"./LegalSkillsProfilePage-Cfuyr1r4.js";import"./button-D5QFqFqd.js";import"./index-H0-EOAh9.js";import"./vendor-icons-BX2-bB2Y.js";import"./input-DeEwMjyU.js";import"./IusPageShell-_KnMXfS8.js";import"./api-CxwxUWSo.js";function l(){return r.jsx(o,{})}export{l as PracticeProfilePage};